gym/requirements.txt
flask==2.0.1
opencv-python==4.7.0.72
mediapipe==0.10.5
numpy==1.23.5
flask-pymongo==2.3.0
bcrypt==4.0.1
python-dotenv==0.19.0
Pillow==9.3.0
tensorflow==2.13.0
Werkzeug==2.0.1
